+++ integration/container/run_linux_test.go
@@ -274,21 +274,26 @@ func TestMacAddressIsAppliedToMainNetworkWithShortID(t *testing.T) {
 	d.StartWithBusybox(ctx, t)
 	defer d.Stop(t)
 
-	apiClient, err := client.NewClientWithOpts(client.FromEnv, client.WithVersion("1.43"))
-	assert.NilError(t, err)
+	for _, tc := range []string{"1.44", "1.52"} {
+		t.Run("api version "+tc, func(t *testing.T) {
+			apiClient, err := client.NewClientWithOpts(client.FromEnv, client.WithVersion(tc))
+			assert.NilError(t, err)
 
-	n := net.CreateNoError(ctx, t, apiClient, "testnet", net.WithIPAM("192.168.101.0/24", "192.168.101.1"))
+			n := net.CreateNoError(ctx, t, apiClient, "testnet", net.WithIPAM("192.168.101.0/24", "192.168.101.1"))
+			defer net.RemoveNoError(ctx, t, apiClient, n)
 
-	cid := container.Run(ctx, t, apiClient,
-		container.WithImage("busybox:latest"),
-		container.WithCmd("/bin/sleep", "infinity"),
-		container.WithStopSignal("SIGKILL"),
-		container.WithNetworkMode(n[:10]),
-		container.WithContainerWideMacAddress("02:42:08:26:a9:55"))
-	defer container.Remove(ctx, t, apiClient, cid, client.ContainerRemoveOptions{Force: true})
+			cid := container.Run(ctx, t, apiClient,
+				container.WithImage("busybox:latest"),
+				container.WithCmd("/bin/sleep", "infinity"),
+				container.WithStopSignal("SIGKILL"),
+				container.WithNetworkMode(n[:10]),
+				container.WithMacAddress("testnet", "02:42:08:26:a9:55"))
+			defer container.Remove(ctx, t, apiClient, cid, client.ContainerRemoveOptions{Force: true})
 
-	c := container.Inspect(ctx, t, apiClient, cid)
-	assert.Equal(t, c.NetworkSettings.Networks["testnet"].MacAddress, "02:42:08:26:a9:55")
+			c := container.Inspect(ctx, t, apiClient, cid)
+			assert.Equal(t, c.NetworkSettings.Networks["testnet"].MacAddress, "02:42:08:26:a9:55")
+		})
+	}
 }
 
 func TestStaticIPOutsideSubpool(t *testing.T) {