+++ integration/networking/mac_addr_test.go
@@ -1,7 +1,6 @@
 package networking
 
 import (
-	"net/netip"
 	"testing"
 
 	"github.com/moby/moby/client"
@@ -154,7 +153,6 @@ func TestInspectCfgdMAC(t *testing.T) {
 		name       string
 		desiredMAC string
 		netName    string
-		ctrWide    bool
 	}{
 		{
 			name:    "generated address default bridge",
@@ -174,12 +172,6 @@ func TestInspectCfgdMAC(t *testing.T) {
 			desiredMAC: "02:42:ac:11:00:42",
 			netName:    "testnet",
 		},
-		{
-			name:       "ctr-wide address default bridge",
-			desiredMAC: "02:42:ac:11:00:42",
-			netName:    "bridge",
-			ctrWide:    true,
-		},
 	}
 
 	for _, tc := range testcases {
@@ -187,9 +179,6 @@ func TestInspectCfgdMAC(t *testing.T) {
 			ctx := testutil.StartSpan(ctx, t)
 
 			var copts []client.Opt
-			if tc.ctrWide {
-				copts = append(copts, client.WithVersion("1.43"))
-			}
 			c := d.NewClientT(t, copts...)
 			defer c.Close()
 
@@ -214,11 +203,7 @@ func TestInspectCfgdMAC(t *testing.T) {
 				opts = append(opts, container.WithNetworkMode(tc.netName))
 			}
 			if tc.desiredMAC != "" {
-				if tc.ctrWide {
-					opts = append(opts, container.WithContainerWideMacAddress(tc.desiredMAC))
-				} else {
-					opts = append(opts, container.WithMacAddress(tc.netName, tc.desiredMAC))
-				}
+				opts = append(opts, container.WithMacAddress(tc.netName, tc.desiredMAC))
 			}
 			id := container.Create(ctx, t, c, opts...)
 			defer c.ContainerRemove(ctx, id, client.ContainerRemoveOptions{
@@ -226,57 +211,8 @@ func TestInspectCfgdMAC(t *testing.T) {
 			})
 
 			inspect := container.Inspect(ctx, t, c, ctrName)
-			configMAC := inspect.Config.MacAddress //nolint:staticcheck // ignore SA1019: field is deprecated, but still used on API < v1.44.
-			assert.Check(t, is.DeepEqual(configMAC, tc.desiredMAC))
+			containerMac := inspect.NetworkSettings.Networks[tc.netName].MacAddress
+			assert.Check(t, is.DeepEqual(containerMac, tc.desiredMAC))
 		})
 	}
-}
-
-// Regression test for https://github.com/moby/moby/issues/47441
-// Migration of a container-wide MAC address to the new per-endpoint setting,
-// where NetworkMode uses network id, and the key in endpoint settings is the
-// network name.
-func TestWatchtowerCreate(t *testing.T) {
-	skip.If(t, testEnv.DaemonInfo.OSType == "windows", "no macvlan")
-
-	ctx := setupTest(t)
-
-	d := daemon.New(t)
-	d.StartWithBusybox(ctx, t)
-	defer d.Stop(t)
-
-	c := d.NewClientT(t, client.WithVersion("1.25"))
-	defer c.Close()
-
-	// Create a "/29" network, with a single address in iprange for IPAM to
-	// allocate, but no gateway address. So, the gateway will get the single
-	// free address. It'll only be possible to start a container by explicitly
-	// assigning an address.
-	const netName = "wtmvl"
-	netId := network.CreateNoError(ctx, t, c, netName,
-		network.WithIPAMRange("172.30.0.0/29", "172.30.0.1/32", ""),
-		network.WithDriver("macvlan"),
-	)
-	defer network.RemoveNoError(ctx, t, c, netName)
-
-	// Start a container, using the network's id in NetworkMode but its name
-	// in EndpointsConfig. (The container-wide MAC address must be merged with
-	// the endpoint config containing the preferred IP address, but the names
-	// don't match.)
-	const ctrName = "ctr1"
-	const ctrIP = "172.30.0.2"
-	const ctrMAC = "02:42:ac:11:00:42"
-	id := container.Run(ctx, t, c,
-		container.WithName(ctrName),
-		container.WithNetworkMode(netId),
-		container.WithContainerWideMacAddress(ctrMAC),
-		container.WithIPv4(netName, ctrIP),
-	)
-	defer c.ContainerRemove(ctx, id, client.ContainerRemoveOptions{Force: true})
-
-	// Check that the container got the expected addresses.
-	inspect := container.Inspect(ctx, t, c, ctrName)
-	netSettings := inspect.NetworkSettings.Networks[netName]
-	assert.Check(t, is.Equal(netSettings.IPAddress, netip.MustParseAddr(ctrIP)))
-	assert.Check(t, is.Equal(netSettings.MacAddress, ctrMAC))
 }